more server addditions and whatnot...
