--- src/main/java/chProgram/Application.java
@@ -1,12 +1,0 @@
-package chProgram;
-
-
-class Application{
-
-
-    public static void main(String[] args) {
-        System.out.println("hello");
-    }
-
-
-}
