+++ src/main/java/chView/CheckersBoard.java
@@ -1,0 +1,4 @@
+package chView;
+
+public class CheckersBoard {
+}
