--- src/main/java/chView/ChessBoard.java
@@ -1,4 +1,0 @@
-package chView;
-
-public class ChessBoard {
-}
