+++ src/main/java/chViewModel/CheckerBoard.java
@@ -1,0 +1,4 @@
+package chViewModel;
+
+public class CheckerBoard {
+}
