--- src/main/java/chViewModel/ChessBoard.java
@@ -1,4 +1,0 @@
-package chViewModel;
-
-public class ChessBoard {
-}
